factor: TIL, ErrorCode 리펙토링

+++ src/main/java/org/example/autoreview/domain/tilpost/service/TILPostService.java
@@ -11,12 +11,14 @@ import org.example.autoreview.domain.tilpost.dto.response.TILPostResponseDto;
 import org.example.autoreview.domain.tilpost.entity.TILPost;
 import org.example.autoreview.domain.tilpost.entity.TILPostRepository;
 import org.example.autoreview.global.exception.errorcode.ErrorCode;
+import org.example.autoreview.global.exception.sub_exceptions.BadRequestException;
 import org.example.autoreview.global.exception.sub_exceptions.NotFoundException;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -47,7 +49,7 @@ public class TILPostService {
     @Transactional(readOnly = true)
     public TILPostResponseDto findById(Long id){
         TILPost tilPost = tilPostRepository.findById(id).orElseThrow(
-                () -> new NotFoundException(ErrorCode.POST_NOT_FOUND)
+                () -> new NotFoundException(ErrorCode.NOT_FOUND_POST)
         );
         return new TILPostResponseDto(tilPost);
     }
@@ -57,7 +59,7 @@ public class TILPostService {
         Long id = requestDto.getId();
 
         TILPost tilPost = tilPostRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException(ErrorCode.POST_NOT_FOUND));
+                .orElseThrow(() -> new NotFoundException(ErrorCode.NOT_FOUND_POST));
 
         tilPost.update(requestDto);
 
@@ -65,9 +67,14 @@ public class TILPostService {
     }
 
     @Transactional
-    public Long delete(Long id){
+    public Long delete(Long id, String email){
         TILPost tilPost = tilPostRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException(ErrorCode.POST_NOT_FOUND));
+                .orElseThrow(() -> new NotFoundException(ErrorCode.NOT_FOUND_POST));
+
+        Member member = tilPost.getMember();
+        if(!Objects.equals(member.getEmail(), email)) {
+            throw new BadRequestException(ErrorCode.UNMATCHED_EMAIL);
+        }
 
         tilPostRepository.delete(tilPost);
         return id;
